requirements.txt
pydantic
fastapi
ipython
PyNaCl
base58
python-jose[cryptography]
redis
pyjwt
pusher
uvicorn
solana
anchorpy
pusher_push_notifications
python-telegram-bot
ipython-profile
requests
